+++ gui_treedb/src/c_treedb_picker.js
@@ -18,7 +18,7 @@
  *          All Rights Reserved.
  ***********************************************************************/
 import {
-    SDATA, SDATA_END, data_type_t, event_flag_t,
+    SDATA, SDATA_END, data_type_t,
     gclass_create, log_error,
     gobj_read_attr, gobj_write_attr,
     gobj_subscribe_event,
@@ -138,7 +138,7 @@ function build_ui(gobj)
     priv.$body = $body;
 
     let $manage = createElement2(
-        ["button", {class: "button is-small", id: "ytreedb-manage-btn",
+        ["button", {class: "button is-small PICKER_MANAGE",
                     i18n: "manage connections"}, "Manage connections"]);
     $manage.addEventListener("click", () => {
         /*  hash routing: jump to the Settings page.  */